Added security

+++ todo/todo.py
@@ -11,7 +11,7 @@ bp = Blueprint('todo', __name__)
 @login_required
 def index():
     _, c = get_db()
-    c.execute('SELECT t.id, t.description, u.username, t.completed, t.created_at FROM todo t JOIN user u ON t.created_by = u.id ORDER BY created_at ASC')
+    c.execute('SELECT t.id, t.description, u.username, t.completed, t.created_at FROM todo t JOIN user u ON t.created_by = u.id WHERE created_by = %s ORDER BY created_at ASC', (g.user['id'],))
     todos = c.fetchall()
     return render_template('todo/index.html', todos = todos)
 
@@ -50,7 +50,7 @@ def update(id):
         if error is not None: flash(error)
         else: 
             db, c = get_db()
-            c.execute('UPDATE todo SET description = %s, completed = %s WHERE id = %s', (description, completed, id))
+            c.execute('UPDATE todo SET description = %s, completed = %s WHERE id = %s AND created_by = %s', (description, completed, id, g.user['id']))
             db.commit()
             return redirect(url_for('todo.index'))
 
@@ -60,6 +60,6 @@ def update(id):
 @login_required
 def delete(id):
     db, c = get_db()
-    c.execute('DELETE FROM todo WHERE id = %s', (id,))
+    c.execute('DELETE FROM todo WHERE id = %s AND created_by = %s', (id, g.user['id']))
     db.commit()
     return redirect(url_for('todo.index'))
